+++ src/main/java/com/uniqueAuction/web/user/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.uniqueAuction.web.user.dto;
+package com.uniqueAuction.web.user.request;
 
 import lombok.Getter;
 import lombok.ToString;
